Keep shared tpyes file in sync
